make stop.sh do nothing in case of missing pidfile

+++ stop.sh
@@ -1,4 +1,6 @@
 #!/bin/bash
 cd /webapps/chess
-kill -15 $(cat chess.pid)
-rm chess.pid
+if [ -f chess.pid ]; then
+  kill -15 $(cat chess.pid)
+  rm chess.pid
+fi
